requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0

# File Processing
PyMuPDF==1.23.5
pdfplumber==0.10.3
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.1

# Database & Vector Store
# psycopg2-binary==2.9.7  # DISABLED FOR PHASE 1 - PostgreSQL not used
pymilvus==2.3.1
faiss-cpu==1.7.4

# LLM & AI
groq==0.4.1
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.42.3
torch>=2.0.0
sentence-transformers==3.0.1

# Email APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
microsoft-graph-core==0.2.2
requests==2.31.0

# Text Processing
beautifulsoup4==4.12.2
email-validator==2.0.0
textstat==0.7.3
nltk==3.8.1

# Utilities
geopy==2.3.0
python-dateutil==2.8.2
hashlib2==1.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
